add obstacle avoid action class
