+++ local_waypoint_server/include/local_waypoint_server/actions/follow_path_action.hpp
@@ -12,22 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef LOCAL_WAYPOINT_SERVER__OBSTACLE_AVOID_ACTION_HPP_
-#define LOCAL_WAYPOINT_SERVER__OBSTACLE_AVOID_ACTION_HPP_
-
-#include <string>
-#include <memory>
-#include "behavior_tree_action_builder/action_node.hpp"
-#include "behavior_tree_action_builder/register_nodes.hpp"
-
-namespace local_waypoint_server
-{
-class ObstacleAvoidAction : public behavior_tree_action_builder::ActionNode
-{
-public:
-  explicit ObstacleAvoidAction(const std::string & name, const BT::NodeConfiguration & config);
-  BT::NodeStatus tick() override;
-};
-}
-
-#endif  // LOCAL_WAYPOINT_SERVER__OBSTACLE_AVOID_ACTION_HPP_
+#ifndef LOCAL_WAYPOINT_SERVER__ACTIONS__FOLLOW_PATH_ACTION_HPP_
+#define LOCAL_WAYPOINT_SERVER__ACTIONS__FOLLOW_PATH_ACTION_HPP_
+#endif  // LOCAL_WAYPOINT_SERVER__ACTIONS__FOLLOW_PATH_ACTION_HPP_
